Second version Dockerfile
The Dockerfile installs Gazebo along with its dependencies on a docker, and sets up X11 so that graphic applications can be shown on the host machine. Also installs xeyes (for debugging and testing the x11 sever and connection) among other utilities. Sets the container to run with the necessary flags for accessing the host's x11 server